requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
supabase==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
anthropic==0.7.8
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1